+++ RacingAidWpf/View/TelemetryOverlay.xaml
@@ -1,11 +1,11 @@
-﻿<overlayManagement:Overlay x:Class="RacingAidWpf.View.TelemetryOverlay"
+﻿<overlays:Overlay x:Class="RacingAidWpf.View.TelemetryOverlay"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:converters="clr-namespace:RacingAidWpf.Converters"
         xmlns:viewModel="clr-namespace:RacingAidWpf.ViewModel"
-        xmlns:overlayManagement="clr-namespace:RacingAidWpf.OverlayManagement"
+        xmlns:overlays="clr-namespace:RacingAidWpf.Overlays"
         mc:Ignorable="d"
         d:DataContext="{d:DesignInstance viewModel:TelemetryOverlayViewModel}"
         Title="Telemetry"
@@ -14,13 +14,13 @@
         Topmost="True"
         AllowsTransparency="True"
         MouseLeftButtonDown="OnMouseLeftButtonDown">
-    <overlayManagement:Overlay.Resources>
+    <overlays:Overlay.Resources>
         <converters:SpeedConverter x:Key="SpeedConverter"/>
         <converters:GearConverter x:Key="GearConverter"/>
-    </overlayManagement:Overlay.Resources>
-    <overlayManagement:Overlay.Background>
+    </overlays:Overlay.Resources>
+    <overlays:Overlay.Background>
         <SolidColorBrush Color="Transparent"/>
-    </overlayManagement:Overlay.Background>
+    </overlays:Overlay.Background>
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition/>
@@ -54,4 +54,4 @@
             <TextBlock Grid.Row="1" Style="{StaticResource DefaultTextBlock}" TextAlignment="Center" FontWeight="ExtraBold" FontSize="40" Foreground="Goldenrod" Text="{Binding Gear, Converter={StaticResource GearConverter}}"/>
         </Grid>
     </Grid>
-</overlayManagement:Overlay>
+</overlays:Overlay>
