+++ RacingAidWpf/View/TimesheetOverlay.xaml
@@ -1,11 +1,11 @@
-﻿<overlayManagement:Overlay x:Class="RacingAidWpf.View.TimesheetOverlay"
+﻿<overlays:Overlay x:Class="RacingAidWpf.View.TimesheetOverlay"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:viewModel="clr-namespace:RacingAidWpf.ViewModel"
         xmlns:converters="clr-namespace:RacingAidWpf.Converters"
-        xmlns:overlayManagement="clr-namespace:RacingAidWpf.OverlayManagement"
+        xmlns:overlays="clr-namespace:RacingAidWpf.Overlays"
         mc:Ignorable="d"
         d:DataContext="{d:DesignInstance viewModel:TimesheetOverlayViewModel}"
         Title="Timesheet"
@@ -16,14 +16,14 @@
         Topmost="True"
         AllowsTransparency="True"
         MouseLeftButtonDown="OnMouseLeftButtonDown">
-    <overlayManagement:Overlay.Background>
+    <overlays:Overlay.Background>
         <SolidColorBrush Color="Transparent"/>
-    </overlayManagement:Overlay.Background>
-    <overlayManagement:Overlay.Resources>
+    </overlays:Overlay.Background>
+    <overlays:Overlay.Resources>
         <converters:LapTimeConverter x:Key="LapTimeConverter"/>
         <converters:GapConverter x:Key="GapConverter"/>
         <converters:PositionConverter x:Key="PositionConverter"/>
-    </overlayManagement:Overlay.Resources>
+    </overlays:Overlay.Resources>
     <Grid>
         
         <Canvas Style="{StaticResource DefaultCanvasStyle}"/>
@@ -56,4 +56,4 @@
             </DataGrid.RowStyle>
         </DataGrid>
     </Grid>
-</overlayManagement:Overlay>
+</overlays:Overlay>
